README.md improvements
Add example of crontab usage.

Fix typo in one of the example URLs.

Change host name in examples from localhost:8080 to
"nights-watch-server".

Add section explaining how to monitor Night's Watch itself.

Add section discussing alternatives to Night's Watch.

Add more to-dos.

Delete PyPI to-do since I'm hoping to do that today and it would be
silly for that to be in README.me in the version that gets released to
PyPI.