Send response data through wsgiref in chunks rather than all at once
If you give wsgiref a single, huge response blob to send, it sends the
data to the socket in a single call to write(), which isn't guaranteed
to send the whole block of data, and wsgiref doesn't check if all the
data was sent and send the rest if it wasn't. This is arguably a bug
in wsgiref, but fixing that will take time and effort, so we work
around it here.