Defer use of pypandoc until after setuptools makes it available
You can't just put `from pypandoc import convert` at the top of your
setup.py and then put `description=convert("README.md", "rst")` in
your `setup()` invocation, because even if you list list `pypandoc` in
`setup_requires`, it won't be interpreted and installed until after
the keyword argument values of the `setup()` invocation have been
evaluated. Therefore, we define a `lazy_convert` class which
impersonates a string but doesn't actually import or use `pypandoc`
until the value of the string is needed. This defers the use of
`pypandoc` until after setuptools has figured out that it is needed
and made it available.