Fetch canaries from store to return after modifying them
After modifying a canary in the store, fetch the stored canary to
return, rather than returning the dict that we passed into the
store.

This is necessary, e.g., because MongoDB rounds datetimes to the
nearest millisecond. When we store a datetime with microsecond
resolution, what gets stored is something with only millisecond
resolution, and that's what we want to return to the caller.