Don't log redundant "Setting alarm for ..." messages
Out of paranoia, Coal Mine resets the alarm for the next upcoming
canary that's due every time a canary is triggered, rather than
assuming that the alarm is intact from the previous time it was set
(if it's still the same canary that's next due).

That's a cheap, reasonable bit of paranoia, but it comes with it a lot
of unnecessary log noise about setting the same alarm over and over
again.

Therefore, we now only log the alarm we're setting when it's different
from the alarm that was set before.