Refactor CronTabSchedule for performance
Newer versions of the third-party packages that Coal Mine uses have
introduced significant performance degradations which caused Coal Mine
to perform much more poorly. This would have been unlikely to be
noticeable in production, but it was making the unit tests several
times slower than they were before. I therefore took the time to
refactor the code to make it smarter and perform better, which wasn't
necessary before, i.e., I got away with the stupid approach before
because it was fast enough, but now I need to be smarter because the
code we're calling is slower.

Having complete code coverage was a huge help in this because it
gave me a way to quickly, accurately confirm that my significant
internal code changes continued to function as expected.