Support env vars instead of INI file; support MongoDB URI
Allow the server to be configured using environment variables rather
than the coal-mine.ini file, and allow the MongoDB connection to be
configured using a MongoDB URI instead of separate host name,
database, username, and password parameters.

We need to add the `srv` extra to the `pymongo` package in
`requirements.txt` for MongoDB URI parsing to work.