Upgrade to current PyPI packages, use package version ranges
Use current versions of dependent PyPI packages.

Use package version ranges in the requirements files instead of
individual package versions, to maximize the use of system-installed
packages and minimize conflicts with other packages. Moving forward we
will attempt to preserve compatibility with old versions of dependent
packages for as long as possible to maximize the environments in which
Coal Mine can be installed without pulling in unnecessary additional
dependent packages.

Bump version number to 0.6.1.